requirements.txt
streamlit
tensorflow
numpy
Pillow
matplotlib
seaborn
scikit-learn
opencv-python